Removed prints from Gyro.h & improved loop timing
There is still a latent bug where the printMaze() function causes 180
degree turns to spaz out. I think it has something to do with inefficient
string construction or Serial printing. But I found out that when the
`printMaze()` function is uncommented, the gyroscope reads a very high
angular speed, which causes the gyroscopes current angle to jump to a
huge value. This causes the PID controller to go crazy and the robot spins
around a bunch.